+++ docs_src/src/changes.md
@@ -2,6 +2,7 @@
 * The bestiary works once again in the beta and the new stats are supported.
 * Added support for the mods menu in settings.
 * Initial support for BaseLib has been implemented. All mod config settings should now be supported, however custom control types added by mods will most likely not work.
+* Hotkeys have been added to quickly read the card count and cards in each card pile. Alt a for draw, alt s for discard, alt d for deck, alt f for exhaust, and alt h for your hand.
 
 ## V1.1.8
 * Fixed a bug introduced in the latest beta where you could not return from the top bar to the rest of the game UI. Again.